addes DeletedRows command and unit tests

+++ src/main/java/controller/commands/DropTableCommand.java
@@ -2,6 +2,7 @@ package controller.commands;
 
 import java.util.List;
 
+import controller.exceptions.ControllerException;
 import model.DbOperations;
 import view.*;
 import view.view.View;
@@ -10,6 +11,8 @@ import view.view.View;
  * Created by stas on 10/29/17.
  */
 public class DropTableCommand implements Command<String> {
+
+    private static final int ALLOWED_PARAMS_SIZE = 1;
 
     private DbOperations dbOperations;
     private View view;
@@ -21,6 +24,9 @@ public class DropTableCommand implements Command<String> {
 
     @Override
     public void execute(List<String> parameters) {
+        if (parameters.size() != ALLOWED_PARAMS_SIZE) {
+            throw new ControllerException("Should be one parameter tableName");
+        }
         final String tableName = parameters.get(0);
         dbOperations.dropTable(tableName);
         view.write(String.format("Table %s is dropped", tableName));
